Protocol changes and allow server plugins to mute the next chat message
Idea: Canceling the chat message on the proxy ensures that chat messages
that are supposed to be interpreted as text input are also not handled by
chat plugins on the proxy, such as being forwarded to other players.

However, this has not yet been tested much, and there might be a remaining
severe issue with this approach: During brief testing on MC 1.19.3, due to
the signed chat additions in Minecraft, it appears that players might get
kicked due to their chat packet missing or having an invalid signature if
they have been temporarily muted and then try to send their next message
after having been unmuted again. Further investigation is necessary.

Other changes:
* Messages are now represented by actual classes. And there have been some
  changes to how these messages are serialized and deserialized: Although
  currently not yet necessary because each channel is used by only one type
  of message, these messages also serialize their type id now. This allows
  for a more generic way of how we serialize and deserialize messages, and
  will be useful should we ever need to send more than one type of message
  via the same channel.
* Add additional debug output regarding the involved thread when receiving
  plugin messages or handling cancelled chat events.
* Extend the Readme.
* Bump the version number.